% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_to_pseudobulk}
\alias{aggregate_to_pseudobulk}
\title{Aggregate counts to pseudobulk}
\usage{
aggregate_to_pseudobulk(
  count_matrix,
  metadata,
  aggregate_by,
  sep = ".",
  format = c("list", "sce")
)
}
\arguments{
\item{count_matrix}{Numeric matrix-like object with features in rows and cells
in columns.}

\item{metadata}{Data frame-like object with one row per cell.}

\item{aggregate_by}{Character vector naming metadata columns used to define
pseudobulk samples.}

\item{sep}{Character string separating metadata values in pseudobulk names.}

\item{format}{EIther list or sce. If list will return a SimpleList with the count matrix and metadata,
and if sce then returns a SingleCellExperiment with first assay being the counts and metadata as colData.}
}
\value{
A list with:
\describe{
  \item{\code{counts}}{Feature-by-pseudobulk matrix of summed counts.}
  \item{\code{metadata}}{Pseudobulk metadata including the number of cells.}
}
}
\description{
Aggregates columns of a count matrix by combinations of metadata variables,
returning summed counts and pseudobulk-level metadata.
}
\examples{
sce <- mockSCE()
sce$group <- rep(c("A", "B"), ncol(sce) / 2)
sce$group2 <- rep(c("A", "B", "C", "D"), each = ncol(sce) / 4)

pb <- aggregate_to_pseudobulk(
  count_matrix = SummarizedExperiment::assay(sce),
  metadata = SummarizedExperiment::colData(sce),
  aggregate_by = c("group", "group2")
)

}
