% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{create_signatures}
\alias{create_signatures}
\title{Create group-specific gene signatures based on pairwise differential analysis results}
\usage{
create_signatures(
  x,
  n = Inf,
  min_prop = 1,
  exclude_groups = NULL,
  signif_column = "adj.P.Val",
  signif_threshold = 0.05,
  effect_column = "logFC",
  effect_threshold = 0,
  rank_column = "P.Value",
  rank_method = "increasing",
  gene_column = "Gene"
)
}
\arguments{
\item{x}{List with DE results from \code{de_limma} run with \code{mode="pairwise"}}

\item{n}{The maximum number of top signature genes per group to return}

\item{min_prop}{A number greater zero and smaller or equal than 1. Defines the stringency of the signature. It is the proportion
of groups that a gene needs to be overexpressed to. A value of 1 means a gene must be overexpressed versus all other groups. A value
of 0.75 means overexpressed versus 75\% of other groups.}

\item{exclude_groups}{Vector with group names to exclude from signature creation. Can be helpful if a group is expected to be transitional between two other groups
so one would not expect unique genes defining this group. Can help returning more genes for the remaining groups.}

\item{signif_column}{Name of the significance column in \code{x}.}

\item{signif_threshold}{Threshold for significance.}

\item{effect_column}{Name of the effect column in \code{x}.}

\item{effect_threshold}{Threshold for effect size \(below this is "underexpressed", above is "overexpressed"\)}

\item{rank_column}{Name of the column in \code{x} to use for ranking.}

\item{rank_method}{Either "decreasing" or "increasing" ranking direction}

\item{gene_column}{Name of the gene column in \code{x}.}
}
\description{
Create group-specific gene signatures based on pairwise differential analysis results
}
\examples{
library(SingleCellExperiment)
set.seed(1)
sce <- mockSCE(ncells = 1000, ngenes = 1000, nspikes = 0)
sce$group <- rep(LETTERS[1:4], each = 250)
sce$donor <- rep(LETTERS[1:4], 250)
sizeFactors(sce) <- rnorm(ncol(sce), mean = 1, sd = .2)
res_pairwise <- de_limma(x = sce, aggregate_by = c("group", "donor"),
                         main_covariate = "group", other_covariates = c("donor"))
create_signatures <- create_signatures(x = res_pairwise, signif_threshold = .95)
create_signatures

}
\author{
Alexander Bender
}
