% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_contrasts_average}
\alias{.make_contrasts_average}
\title{Makes all averaging contrasts based on an intercept-less design}
\usage{
.make_contrasts_average(u, design)
}
\description{
Makes all averaging contrasts based on an intercept-less design
}
\keyword{internal}
