% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.make_contrasts_pairwise}
\alias{.make_contrasts_pairwise}
\title{Makes all pairwise contrasts based on an intercept-less design}
\usage{
.make_contrasts_pairwise(u, design, delim)
}
\description{
Makes all pairwise contrasts based on an intercept-less design
}
\keyword{internal}
