% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rankmatrix}
\alias{.rankmatrix}
\title{Modified from Kolde et al in \code{RobustRankAggreg::rankMatrix()}}
\usage{
.rankmatrix(x)
}
\arguments{
\item{x}{list with ranked elements}
}
\description{
Modified from Kolde et al in \code{RobustRankAggreg::rankMatrix()}
}
