% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{rank_degs}
\alias{rank_degs}
\title{Internal function to filter and rank DEG results}
\usage{
rank_degs(
  res,
  delim,
  signif_column,
  signif_threshold,
  effect_column,
  effect_threshold,
  rank_column,
  rank_method,
  gene_column,
  min_pct
)
}
\arguments{
\item{res}{The output of \code{de_limma} run with \code{mode="pairwise"}}

\item{signif_column}{Name of the significance column in \code{x}.}

\item{signif_threshold}{Threshold for significance.}

\item{effect_column}{Name of the effect column in \code{x}.}

\item{effect_threshold}{Threshold for effect size \(below this is "underexpressed", above is "overexpressed"\)}

\item{rank_column}{Name of the column in \code{x} to use for ranking.}

\item{rank_method}{Either "decreasing" or "increasing" ranking direction}

\item{gene_column}{Name of the gene column in \code{x}.}
}
\description{
Internal function to filter and rank DEG results
}
\author{
Alexander Bender
}
\keyword{internal}
