% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{ranks2signatures}
\alias{ranks2signatures}
\title{Convert ranks to signatures}
\usage{
ranks2signatures(ranked, min_prop, n, exclude_groups)
}
\arguments{
\item{ranked}{The output of \code{rank_degs}}

\item{min_prop}{A number greater zero and smaller or equal than 1. Defines the stringency of the signature. It is the proportion
of groups that a gene needs to be overexpressed to. A value of 1 means a gene must be overexpressed versus all other groups. A value
of 0.75 means overexpressed versus 75\% of other groups.}

\item{n}{The maximum number of top signature genes per group to return}

\item{exclude_groups}{Vector with group names to exclude from signature creation. Can be helpful if a group is expected to be transitional between two other groups
so one would not expect unique genes defining this group. Can help returning more genes for the remaining groups.}
}
\description{
Convert ranks to signatures
}
\author{
Alexander Bender
}
\keyword{internal}
