\name{siggenes-internal}
\alias{adjust.for.mt}
\alias{build.dperm}
\alias{cat.stat}
\alias{d.null}
\alias{d.stat}
\alias{pairt.cl.transform}
\alias{pairt.samp}
\alias{pairt.samp.transform}
\alias{Rfold.cal}
\alias{setup.mat.samp}
\alias{stats.cal}
\alias{wilc.stat}
\alias{ebam.old}
\alias{ebam.wilc.old}
\alias{find.a0.old}
\alias{neglogLik.repeat}
\alias{q.value.old}
\alias{q.value.wilc}
\alias{R.fold.old}
\alias{ratio.est}
\alias{rs.cal}
\alias{sam.sampler}
\alias{xy.cal}
\title{Internal siggenes functions}
\description{Internal siggenes functions.}
\details{
   These functions are not meant to be directly called by the user.
}
\author{Holger Schwender, \email{holger.schw@gmx.de}}
\keyword{internal}{siggenes-internal}
