% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_testing.R
\name{de_limma}
\alias{de_limma}
\title{Automated differential expression via the limma package}
\usage{
de_limma(
  x,
  use_assay = "counts",
  aggregate_by = NULL,
  use_directly = FALSE,
  limma_method = c("trend", "voom"),
  prefilter_method = c("percent_expressed", "filterByExpr", "none"),
  mode = c("pairwise", "average"),
  main_covariate,
  other_covariates = NULL,
  block = NULL,
  min_pct = 0,
  min_fc = 1,
  use_weights = FALSE,
  delim = "_vs_",
  return_object = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A `SingleCellExperiment`}

\item{use_assay}{Name of the assay used for testing, "counts" by default}

\item{aggregate_by}{A character vector of column names in `colData(x)` used for pseudobulk aggregation.}

\item{use_directly}{Logical, if `TRUE`, use provided assay directly for testing. Assumes data are prefiltered and on log2 scale.}

\item{limma_method}{Either "trend" or "voom". See details}

\item{prefilter_method}{The prefiltering strategy, see details}

\item{mode}{Comparison mode, either "pairwise" (all-vs-all) or `"average"` (one-vs-avg.of.rest)}

\item{main_covariate}{The main grouping variable in `colData(x)` used for contrasts}

\item{other_covariates}{Optional character vector of additional covariates to include in the model}

\item{block}{Optional blocking variable (e.g., donor ID) from `colData(x)` for limma's duplicate correlation}

\item{min_pct}{Numeric between 0 and 100, minimum percent expression per group required to keep a gene (used with "percent_expressed" method).}

\item{min_fc}{Minimum absolute fold change threshold (non-log scale) used in `limma::treat`}

\item{use_weights}{Logical, if `TRUE`, will estimate empirical sample weights}

\item{delim}{Delimiter used to name pairwise contrasts (default: "_vs_").}

\item{return_object}{Logical, if `TRUE`, return `DGEList` object and design/contrast info as attributes. Will be the pseudobulked one if that option was chosen}

\item{verbose}{Logical, if `TRUE`, print progress messages}
}
\description{
Automated differential expression via the limma package
}
\examples{

# Testing on pseudobulk level with limma-voom
library(SingleCellExperiment)
set.seed(1)
sce <- mockSCE(ncells = 1000, ngenes = 1000, nspikes = 0)
sce$group <- rep(LETTERS[1:4], each = 250)
sce$donor <- rep(LETTERS[1:4], 250)
sizeFactors(sce) <- rnorm(ncol(sce), mean = 1, sd = .2)

res_pairwise <- de_limma(
  x = sce, aggregate_by = c("group", "donor"), mode = "pairwise",
  main_covariate = "group", other_covariates = c("donor"), limma_method = "trend"
)

}
\author{
Alexander Bender
}
