% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.aggreg_pseudobulk}
\alias{.aggreg_pseudobulk}
\title{Aggregate a SingleCellExperiment assay to pseudobulk sums using only base R}
\usage{
.aggreg_pseudobulk(count_matrix, metadata, aggregate_by, sep = ".")
}
\arguments{
\item{count_matrix}{Matrix of (raw) counts to aggregate}

\item{metadata}{Data.frame with annotations}

\item{aggregate_by}{Character vector indicating which columns of metadata to use}

\item{sep}{String used to delimit colnames in aggregated data.frame}
}
\description{
Aggregate a SingleCellExperiment assay to pseudobulk sums using only base R
}
\examples{
sce <- mockSCE()
sce$group <- rep(c("A", "B"), ncol(sce)/2)
sce$group2 <- rep(c("A", "B", "C", "D"), each = ncol(sce)/4)
pb <- .aggreg_pseudobulk(count_matrix = assay(sce), metadata = colData(sce), aggregate_by = c("group", "group2"), sep = ".")
}
