% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pexpr}
\alias{get_pexpr}
\title{Calculate percentage of expression per group for each gene.}
\usage{
get_pexpr(data, group, threshold = 0, digits = 2)
}
\arguments{
\item{data}{Numeric data.frame, matrix or similar with rows being genes/observations and columns being cells/samples.}

\item{group}{Vector of length \code{ncol(data)} with group information}

\item{threshold}{Consider counts above this threshold as "expressed"}

\item{digits}{Round percentage to this digit}
}
\description{
Calculate percentage of expression per group for each gene.
}
\examples{
ngenes <- 10
nsamples <- 1000
set.seed(1)
data <- matrix(rnbinom(ngenes * nsamples, mu = 0.5, nsamples), ncol = nsamples, byrow = TRUE)
rownames(data) <- paste0("gene", 1:nrow(data))
group <- rep(LETTERS[1:10], each = nsamples / 10)
get_pexpr(data, group)

# Also works on Csparse matrix
data_sparse <- as(data, "CsparseMatrix")
get_pexpr(data_sparse, group)

}
\author{
Alexander Bender
}
