\name{siggenes-internal}
\alias{find.cuts}
\alias{fudge}
\alias{na.replace}
\alias{neglogLik.repeat}
\alias{p0.est}
\alias{q.value.cal}
\alias{q.value.wilc}
\alias{R.fold.cal}
\alias{ratio.est}
\alias{roller.coaster}
\alias{rs.cal}
\alias{sam.fdr}
\alias{sam.ngenes}
\alias{sam.plotter}
\alias{sam.sampler}
\alias{wilc.cal}
\title{Internal siggenes functions}
\description{Internal siggenes functions.}
\details{These functions are not meant to be called directly by the user.}
\author{Holger Schwender, \email{holger.schw@gmx.de}}
\keyword{internal}{siggenes-internal}