\name{siggenes-internal}
\alias{adjust.for.mt}
\alias{build.dperm}
\alias{d.null}
\alias{pairt.cl.transform}
\alias{pairt.samp}
\alias{pairt.samp.transform}
\alias{Rfold.cal}
\alias{setup.mat.samp}
\alias{stats.cal}
\alias{quantiles}
\alias{add.target2href}
\alias{check.chipname}
\alias{col2hex}
\alias{getTD4Affy}
\alias{make.tablecode}
\alias{pretty.mat.fdr}
\alias{formatSAM}
\alias{pretty.mat.sig}
\alias{filterALL}
\alias{compPermStat}
\alias{cat.null}
\alias{cat.null.approx}
\alias{chisqClass}
\alias{computeContCols}
\alias{checkA0}
\alias{checkFUNout}
\alias{checkQuantiles}
\alias{compFailure}
\alias{compFailureMat2}
\alias{compFailureSubset}
\alias{compFalse}
\alias{compNumber}
\alias{compRatio}
\alias{computeRS}
\alias{ebamA0}
\alias{getFailure}
\alias{getSuccesses}
\alias{makeA0mat}
\alias{pi0.est2}
\alias{pi0.est3}
\alias{truncZ}
\alias{finda02html}
\alias{cat.null.approx2}
\alias{na.handling}
\alias{na.replace.cont}
\alias{getQuantiles}
\alias{recodeVal}
\alias{recodeLevel}
\alias{getTD4rs}
\alias{findFDR}
\alias{findNumber}
\alias{checkInitialDelta}


\title{Internal siggenes functions}
\description{Internal siggenes functions.}
\details{
   These functions are not meant to be directly called by the user.
}
\author{Holger Schwender, \email{holger.schw@gmx.de}}
\keyword{internal}{siggenes-internal}
